refactored code to put modal in root
