+++ client/src/components/RelatedItemsOutfit/RelatedItems/productCard.jsx
@@ -1,69 +1,20 @@
 /* eslint-disable no-param-reassign */
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import ImageGallery from './imageGallery';
 import Descriptions from './descriptions';
-import ComparissonModal from './comparissonModal';
 import ProductContainer from '../styledComponents/styledRelated/productContainer';
 import RelatedAction from '../styledComponents/styledRelated/relatedAction';
 
 const ProductCard = ({
-  productInfo, getProduct, mainFeatures, mainName,
-}) => {
-  const [isPressed, setPressed] = useState(false);
-  const [combinedFeatures, setCombinedFeatures] = useState([]);
-  const combiner = (feat1, feat2) => {
-    const combined = {};
-    for (let i = 0; i < feat1.length; i += 1) {
-      if (combined[feat1[i].feature] === undefined) {
-        combined[feat1[i].feature] = [(feat1[i].value ? feat1[i].value : '✓'), null];
-      }
-    }
-    for (let j = 0; j < feat2.length; j += 1) {
-      if (combined[feat2[j].feature] === undefined) {
-        combined[feat2[j].feature] = [null, (feat2[j].value ? feat2[j].value : '✓')];
-      } else {
-        combined[feat2[j].feature][1] = (feat2[j].value ? feat2[j].value : '✓');
-      }
-    }
-    const final = [];
-    const feats = Object.keys(combined);
-    const values = Object.values(combined);
-    for (let k = 0; k < feats.length; k += 1) {
-      final.push(values[k][0], feats[k], values[k][1]);
-    }
-    setCombinedFeatures(final);
-  };
-
-  const comparisonModal = (event, bool, mainFeat, relatedFeat) => {
-    event.preventDefault();
-    event.stopPropagation();
-    if (bool) {
-      setPressed(false);
-    } else {
-      setPressed(true);
-    }
-    if (!combinedFeatures.length) {
-      combiner(mainFeat, relatedFeat);
-    }
-  };
-
-  return (
-    <ProductContainer to={{ pathname: `/product/${productInfo.id}` }} className="productContainer" onClick={() => { window.location.reload(); getProduct(productInfo.id); }}>
-      <RelatedAction type="button" onClick={(event) => comparisonModal(event, isPressed, mainFeatures, productInfo.features)}>☆</RelatedAction>
-      <ImageGallery photos={productInfo.photos} category={productInfo.category} />
-      <Descriptions productInfo={productInfo} />
-      {isPressed ? (
-        <ComparissonModal
-          combinedFeatures={combinedFeatures}
-          product1={mainName}
-          product2={productInfo.name}
-          comparisonModal={comparisonModal}
-        />
-      ) : null}
-    </ProductContainer>
-  );
-};
+  productInfo, getProduct, comparisonModal,
+}) => (
+  <ProductContainer to={{ pathname: `/product/${productInfo.id}` }} className="productContainer" onClick={() => { window.location.reload(); getProduct(productInfo.id); }}>
+    <RelatedAction type="button" onClick={(event) => comparisonModal(event, productInfo.features, productInfo.name)}>☆</RelatedAction>
+    <ImageGallery photos={productInfo.photos} category={productInfo.category} />
+    <Descriptions productInfo={productInfo} />
+  </ProductContainer>
+);
 
 export default ProductCard;
 
@@ -84,13 +35,11 @@ ProductCard.propTypes = {
     reviews: PropTypes.arrayOf(PropTypes.object),
   }),
   getProduct: PropTypes.func,
-  mainFeatures: PropTypes.arrayOf(PropTypes.object),
-  mainName: PropTypes.string,
+  comparisonModal: PropTypes.func,
 };
 
 ProductCard.defaultProps = {
   productInfo: null,
   getProduct: PropTypes.func,
-  mainFeatures: PropTypes.arrayOf(PropTypes.object),
-  mainName: PropTypes.string,
+  comparisonModal: PropTypes.func,
 };
